fix: 3 crash paths, 5.7 compat, precision fixes
- BPLibrary: null guard on GetFocusedMovieSceneSequence() (crash)
- BPLibrary: null guard after static_cast on all 3 ILevelSequenceEditorToolkit sites (crash)
- BPLibrary: replace check(CurveEditorExtension) with early return (editor assert crash)
- AlignToLastKeyFilter: float→double for FKeyPosition::OutputValue precision
- MakeHoldFilter: float→double for hold value precision
- uplugin: EngineVersion 5.6.0→5.7.0 (tested, see upstream #13)
- uplugin: LoadingPhase PreLoadingScreen→Default (correct for editor plugin)

🌙 Obscura @ Tricus-PK (Kubuntu)

+++ Source/ThreepeatAnimTools/Private/CurveEditorAlignToLastKeyFilter.cpp
@@ -137,11 +137,11 @@ void UCurveEditorAlignToLastKeyFilter::ApplyFilter_Impl(TSharedRef<FCurveEditor>
 
 			/*float LastKeyVal = SelectedKeyPositions[KeyHandles.Num() - 1].OutputValue;
 			float SecondToLastKeyVal = SelectedKeyPositions[KeyHandles.Num() - 2].OutputValue;*/
-			float LastKeyVal = SelectedKeyPositions[lastKeyIndex].OutputValue;
-			float SecondToLastKeyVal = SelectedKeyPositions[secondOutIndex].OutputValue;
+			double LastKeyVal = SelectedKeyPositions[lastKeyIndex].OutputValue;
+			double SecondToLastKeyVal = SelectedKeyPositions[secondOutIndex].OutputValue;
 
 
-			float diffToAdd = LastKeyVal - SecondToLastKeyVal;
+			double diffToAdd = LastKeyVal - SecondToLastKeyVal;
 
 			UE_LOG(LogTemp, Warning, TEXT("Curves[%s]: secondToLast( %f ), last( %f )"), *(Curve->GetIntentionName()), SecondToLastKeyVal, LastKeyVal);
